Added emails with gamekey sending(gamekeys is from database, after purchase they will be deleted)

+++ app/Http/Controllers/CartController.php
@@ -4,8 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Game;
+use App\Models\Gamekey;
 use App\Models\Order;
+use App\Models\User;
 use Auth;
+use Mail;
+use App;
+use App\Mail\GamesTopPurchaseEmail;
 
 class CartController extends Controller
 {
@@ -57,6 +62,7 @@ class CartController extends Controller
     {
         $user_id = Auth::user()->id;
         $game_id = Game::where('name', $request->gamename)->first()->id;
+        $keys_count = Gamekey::where("game_id", $game_id)->get()->count();
         
         $result = false;
         
@@ -69,50 +75,59 @@ class CartController extends Controller
         
         if($order != null)
         {
-            $values = $order->toArray(); //I dont know how I can get `count` column in another way
-            
+            //how big we want to change cart
             $increaseCount = 1;
             if(isset($request->changeCount)) //if we`re changing count from cart
             {
                 $increaseCount = $request->changeCount;
             }
             
-            //getting current count, and increasing/decreasing it
-            $newCount = $values['count'] + $increaseCount;
-            
-            if($newCount <= 0) //is user wants to delete order 
+            if($increaseCount + $order->count <= $keys_count)
             {
-                $order->delete();
-                $result = true;
-            }
-            else
-            {
-                $order->update(['count' => $newCount]);
-                $result = true;
-            }
+                $values = $order->toArray(); //I dont know how I can get `count` column in another way
             
-            $orders = Order::where('user_id', $user_id)->get();
-            
-            return view("main.ajax.newcart", [
-                "orders" => $orders,
-                "cartSum" => $this->cartSum($user_id)
-            ])->render();
+                //getting current count, and increasing/decreasing it
+                $newCount = $values['count'] + $increaseCount;
+
+                if($newCount <= 0) //is user wants to delete order 
+                {
+                    $order->delete();
+                    $result = true;
+                }
+                else
+                {
+                    $order->update(['count' => $newCount]);
+                    $result = true;
+                }
+
+                $orders = Order::where('user_id', $user_id)->get();
+
+                return view("main.ajax.newcart", [
+                    "orders" => $orders,
+                    "cartSum" => $this->cartSum($user_id)
+                ])->render();
+            }
+            else return response()->json(false);
         }
         else
         {
-            //creating new order
-            $order = new Order;
+            if($keys_count >= 1) //if we still have enough keys
+            {
+                //creating new order
+                $order = new Order;
+
+                $order->game_id = $game_id;
+                $order->user_id = $user_id;
+                $order->count = 1;
+
+                $order->save();
+
+                $result = true;
+            }
             
-            $order->game_id = $game_id;
-            $order->user_id = $user_id;
-            $order->count = 1;
-            
-            $order->save();
-            
-            $result = true;
+            return response()->json(["Result" => $result]);
         }
         
-        return response()->json($result);
     }
     
     public function purchase()
@@ -120,7 +135,56 @@ class CartController extends Controller
         if(Auth::check()) //in order to avoid some bugs
         {
             $user_id = Auth::user()->id;
-            $orders = Order::where('user_id', $user_id)->delete();
+            $user = User::where("id", $user_id)->first();
+            
+            $orders = Order::where('user_id', $user_id)->get();
+            
+            //for user statistics
+            $total_keys_purchasing = 0;
+            $total_money_wasting = 0;
+            
+            //we will store gamekeys_here
+            $keys = [];
+            
+            //calculating stats and deleting orders
+            foreach($orders as $order)
+            {
+                //stats for user
+                $total_keys_purchasing += $order->count;
+                $total_money_wasting += $order->count * $order->game->price;
+                
+                //stats for game
+                $new_total_keys_purchased = $order->game->purchase_count + $total_keys_purchasing;
+                $order->game->update([
+                    "purchase_count" => $new_total_keys_purchased
+                ]);
+                
+                //adding gamekey to temp array and deleting it
+                $gamekeys = Gamekey::where("game_id", $order->game->id)->get();
+                for($i = 0; $i < $total_keys_purchasing; $i++)
+                {
+                    $current_key = $gamekeys[$i];
+                    array_push($keys, [
+                        "game" => $order->game->name,
+                        "code" => $current_key->code
+                    ]);
+                    $current_key->delete();
+                }
+                
+                $order->delete();
+            }
+            
+            //updating user stats
+            $new_total_keys = $user->keys_purchased + $total_keys_purchasing;
+            $new_total_wasted = $user->total_wasted + $total_money_wasting;
+            
+            $user->update([
+                "keys_purchased" => $new_total_keys,
+                "total_wasted" => $new_total_wasted
+            ]);
+            
+            //sending mail
+            Mail::send(new GamesTopPurchaseEmail($user->email, $keys));
             
             return view("main.ajax.thanks")->render();
         }
